package/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

keras==2.11.0
numba==0.56.4
numpy==1.22.3
scikit-image==0.19.3
tensorflow-intel==2.11.0
